# services/auth-service/target/generated-sources/openapi/src/main/java/com/medicalai/authservice/api/AuthApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.4.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.medicalai.authservice.api;

import com.medicalai.authservice.model.AuthResponse;
import com.medicalai.authservice.model.ErrorResponse;
import com.medicalai.authservice.model.LoginRequest;
import com.medicalai.authservice.model.RegisterRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-05-19T21:34:11.321215300+02:00[Europe/Rome]", comments = "Generator version: 7.4.0")
@Validated
@Tag(name = "Auth", description = "Operazioni di autenticazione")
public interface AuthApi {

    /**
     * POST /auth/login : Autentica un utente e restituisce un token JWT
     *
     * @param loginRequest  (required)
     * @return Login effettuato con successo (status code 200)
     *         or Credenziali non valide (status code 401)
     */
    @Operation(
        operationId = "loginUser",
        summary = "Autentica un utente e restituisce un token JWT",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Login effettuato con successo", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Credenziali non valide", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/login",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthResponse> loginUser(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    );


    /**
     * POST /auth/register : Registra un nuovo utente
     *
     * @param registerRequest  (required)
     * @return Utente registrato con successo (status code 201)
     *         or Dati non validi (es. email già esistente) (status code 400)
     */
    @Operation(
        operationId = "registerUser",
        summary = "Registra un nuovo utente",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Utente registrato con successo", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Dati non validi (es. email già esistente)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/auth/register",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthResponse> registerUser(
        @Parameter(name = "RegisterRequest", description = "", required = true) @Valid @RequestBody RegisterRequest registerRequest
    );


    /**
     * GET /auth/validate : Valida il token JWT
     *
     * @return Token valido (status code 200)
     *         or Token non valido o mancante (status code 401)
     */
    @Operation(
        operationId = "validateToken",
        summary = "Valida il token JWT",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Token valido"),
            @ApiResponse(responseCode = "401", description = "Token non valido o mancante")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/auth/validate"
    )
    
    ResponseEntity<Void> validateToken(
        
    );

}